+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - display, logins, submissions in admin page.
   check post /answers/:n
 - validate if :type "python" found.
+- can see future issues from current `/tasks/all`.
+- recent uploader
+- keep log/konpy.log when `service stop konpy`.
+
+
+## 0.7.4 (2025-05-07)
 
 
 ## 0.7.3 (2025-05-07)